Filter sankey flow by highest frequency paths
I reshaped and refactored the sankey dataframe to get a count of duplicate rows instead of having lots of duplicate rows with a `value` of 1. This enabled me to sort by value and take the top N for charting.

Also changed the palettes for both plots to use the red/yellow cat/dog scheme.